+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/components/comment/commenteditor.tsx","./src/components/comment/commentitem.tsx","./src/components/comment/commentlist.tsx","./src/components/detailplaylist/detailplaylist.styles.ts","./src/components/detailplaylist/detailplaylist.tsx","./src/components/feedlist/feedlist.styles.ts","./src/components/feedlist/feedlist.tsx","./src/components/feedlist/footer.tsx","./src/components/input/input.tsx","./src/components/layout/layout.tsx","./src/components/loadingspinner/loading.module.ts","./src/components/loadingspinner/loading.tsx","./src/components/modal/modal.tsx","./src/components/modal/modalfull.tsx","./src/components/modal/modal.style.ts","./src/components/modal/modalfull.style.ts","./src/components/profile/profile.styles.ts","./src/components/profile/profile.tsx","./src/components/textarea/textarea.tsx","./src/components/toastmessage/toastmessagecontainer.tsx","./src/components/toastmessage/toastmessageitem.tsx","./src/components/toastmessage/toastmessage.module.ts","./src/components/toastmessage/toastmessagecontainer.module.ts","./src/components/viewvideolist/viewvideolist.style.ts","./src/components/viewvideolist/viewvideolist.tsx","./src/components/button/button.styles.ts","./src/components/button/button.tsx","./src/components/header/header-main/header.styled.ts","./src/components/header/header-main/header.tsx","./src/components/header/header-sub/header.styled.ts","./src/components/header/header-sub/headersub.tsx","./src/components/navbar/navbar.styled.ts","./src/components/navbar/navbar.tsx","./src/components/playlist/footer.tsx","./src/components/playlist/playlist.module.ts","./src/components/playlist/playlistskeleton.tsx","./src/components/playlist/playlist.styles.ts","./src/components/playlist/playlist.tsx","./src/hooks/usescrolllock.ts","./src/pages/home.tsx","./src/pages/index.ts","./src/pages/login/index.tsx","./src/pages/myplaylist/index.tsx","./src/pages/mypage/index.tsx","./src/pages/notfound/index.tsx","./src/pages/playlistcreate/index.tsx","./src/pages/profileedit/index.tsx","./src/pages/signup/index.tsx","./src/pages/subscriptions/index.tsx","./src/pages/userprofile/index.tsx","./src/pages/view/index.tsx","./src/providers/toastmessageprovider.tsx","./src/providers/index.tsx","./src/routes/router.tsx","./src/styles/globalstyles.ts","./src/styles/components/comment/comment-editor.style.ts","./src/styles/components/comment/comment-item.style.ts","./src/styles/components/comment/comment-list.style.ts","./src/styles/pages/view/view.style.ts","./src/types/list.ts","./src/types/toastmessage.d.ts","./src/types/view.ts","./src/types/button.ts","./src/types/form.ts","./src/types/modal.ts","./src/types/profile.ts","./vite-env.d.ts"],"version":"5.7.2"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/apis/axiosinstance.ts","./src/components/comment/commenteditor.tsx","./src/components/comment/commentitem.tsx","./src/components/comment/commentlist.tsx","./src/components/detailplaylist/detailplaylist.styles.ts","./src/components/detailplaylist/detailplaylist.tsx","./src/components/feedlist/feedlist.styles.ts","./src/components/feedlist/feedlist.tsx","./src/components/feedlist/footer.tsx","./src/components/input/input.tsx","./src/components/layout/layout.tsx","./src/components/loadingspinner/loading.module.ts","./src/components/loadingspinner/loading.tsx","./src/components/modal/modal.tsx","./src/components/modal/modalfull.tsx","./src/components/modal/modal.style.ts","./src/components/modal/modalfull.style.ts","./src/components/playlist/footer.tsx","./src/components/playlist/playlist.module.ts","./src/components/playlist/playlist.styles.ts","./src/components/playlist/playlistskeleton.tsx","./src/components/playlist/playlist.tsx","./src/components/profile/profile.styles.ts","./src/components/profile/profile.tsx","./src/components/textarea/textarea.tsx","./src/components/toastmessage/toastmessagecontainer.tsx","./src/components/toastmessage/toastmessageitem.tsx","./src/components/toastmessage/toastmessage.module.ts","./src/components/toastmessage/toastmessagecontainer.module.ts","./src/components/viewvideolist/viewvideolist.style.ts","./src/components/viewvideolist/viewvideolist.tsx","./src/components/button/button.styles.ts","./src/components/button/button.tsx","./src/components/header/header-main/header.styled.ts","./src/components/header/header-main/header.tsx","./src/components/header/header-sub/header.styled.ts","./src/components/header/header-sub/headersub.tsx","./src/components/navbar/navbar.styled.ts","./src/components/navbar/navbar.tsx","./src/components/playlistcreate/playlistcreate.styles.ts","./src/components/playlistcreate/playlistinfo.tsx","./src/components/playlistcreate/playlistispublic.tsx","./src/hooks/usescrolllock.ts","./src/hooks/usevideolink.ts","./src/pages/home.tsx","./src/pages/index.ts","./src/pages/login/index.tsx","./src/pages/login/login.styles.ts","./src/pages/mypage/index.tsx","./src/pages/myplaylist/index.tsx","./src/pages/notfound/index.tsx","./src/pages/playlistcreate/index.tsx","./src/pages/profileedit/index.tsx","./src/pages/signup/index.tsx","./src/pages/subscriptions/index.tsx","./src/pages/userprofile/index.tsx","./src/pages/view/index.tsx","./src/providers/toastmessageprovider.tsx","./src/providers/index.tsx","./src/routes/router.tsx","./src/schema/loginschema.ts","./src/styles/globalstyles.ts","./src/styles/components/comment/comment-editor.style.ts","./src/styles/components/comment/comment-item.style.ts","./src/styles/components/comment/comment-list.style.ts","./src/styles/pages/view/view.style.ts","./src/types/list.ts","./src/types/toastmessage.d.ts","./src/types/view.ts","./src/types/button.ts","./src/types/form.ts","./src/types/modal.ts","./src/types/playlistcreate.ts","./src/types/profile.ts","./vite-env.d.ts"],"version":"5.7.2"}
